split ESP8266WiFiClass in different sub classes for better overview.
