+++ libraries/ESP8266WiFi/src/ESP8266WiFiSTA.h
@@ -1,0 +1,101 @@
+/*
+ ESP8266WiFiSTA.h - esp8266 Wifi support.
+ Based on WiFi.h from Ardiono WiFi shield library.
+ Copyright (c) 2011-2014 Arduino.  All right reserved.
+ Modified by Ivan Grokhotkov, December 2014
+ Reworked by Markus Sattler, December 2015
+
+ This library is free software; you can redistribute it and/or
+ modify it under the terms of the GNU Lesser General Public
+ License as published by the Free Software Foundation; either
+ version 2.1 of the License, or (at your option) any later version.
+
+ This library is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ Lesser General Public License for more details.
+
+ You should have received a copy of the GNU Lesser General Public
+ License along with this library; if not, write to the Free Software
+ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+ */
+
+#ifndef ESP8266WIFISTA_H_
+#define ESP8266WIFISTA_H_
+
+
+#include "ESP8266WiFiType.h"
+#include "ESP8266WiFiGeneric.h"
+
+
+class ESP8266WiFiSTAClass {
+        // ----------------------------------------------------------------------------------------------
+        // ---------------------------------------- STA function ----------------------------------------
+        // ----------------------------------------------------------------------------------------------
+
+    public:
+
+        int begin(const char* ssid, const char *passphrase = NULL, int32_t channel = 0, const uint8_t* bssid = NULL);
+        int begin(char* ssid, char *passphrase = NULL, int32_t channel = 0, const uint8_t* bssid = NULL);
+        int begin();
+
+        void config(IPAddress local_ip, IPAddress gateway, IPAddress subnet);
+        void config(IPAddress local_ip, IPAddress gateway, IPAddress subnet, IPAddress dns);
+
+        bool reconnect();
+        int disconnect(bool wifioff = false);
+
+        uint8_t waitForConnectResult();
+
+        // STA network info
+        IPAddress localIP();
+
+        uint8_t * macAddress(uint8_t* mac);
+        String macAddress();
+
+        IPAddress subnetMask();
+        IPAddress gatewayIP();
+        IPAddress dnsIP(uint8_t dns_no = 0);
+
+        String hostname();
+        bool hostname(char* aHostname);
+        bool hostname(const char* aHostname);
+        bool hostname(String aHostname);
+
+        // STA WiFi info
+        wl_status_t status();
+        String SSID() const;
+        String psk() const;
+
+        uint8_t * BSSID();
+        String BSSIDstr();
+
+        int32_t RSSI();
+
+    protected:
+
+      static bool _useStaticIp;
+
+    // ----------------------------------------------------------------------------------------------
+    // ------------------------------------ STA remote configure  -----------------------------------
+    // ----------------------------------------------------------------------------------------------
+
+    public:
+
+        bool beginWPSConfig(void);
+
+        void beginSmartConfig();
+        bool smartConfigDone();
+        void stopSmartConfig();
+
+
+    protected:
+
+        static bool _smartConfigStarted;
+        static bool _smartConfigDone;
+        static void _smartConfigCallback(uint32_t status, void* result);
+
+};
+
+
+#endif /* ESP8266WIFISTA_H_ */
