+++ libraries/ESP8266WiFi/src/ESP8266WiFiType.h
@@ -1,0 +1,52 @@
+/*
+ ESP8266WiFiType.h - esp8266 Wifi support.
+ Copyright (c) 2011-2014 Arduino.  All right reserved.
+ Modified by Ivan Grokhotkov, December 2014
+ Reworked by Markus Sattler, December 2015
+
+ This library is free software; you can redistribute it and/or
+ modify it under the terms of the GNU Lesser General Public
+ License as published by the Free Software Foundation; either
+ version 2.1 of the License, or (at your option) any later version.
+
+ This library is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ Lesser General Public License for more details.
+
+ You should have received a copy of the GNU Lesser General Public
+ License along with this library; if not, write to the Free Software
+ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+ */
+
+
+#ifndef ESP8266WIFITYPE_H_
+#define ESP8266WIFITYPE_H_
+
+#include <queue.h>
+
+#define WIFI_SCAN_RUNNING   (-1)
+#define WIFI_SCAN_FAILED    (-2)
+
+// Note:
+// this enums need to be in sync with the SDK!
+
+typedef enum WiFiMode {
+    WIFI_OFF = 0, WIFI_STA = 1, WIFI_AP = 2, WIFI_AP_STA = 3
+} WiFiMode_t;
+
+typedef enum {
+    WIFI_PHY_MODE_11B = 1, WIFI_PHY_MODE_11G = 2, WIFI_PHY_MODE_11N = 3
+} WiFiPhyMode_t;
+
+typedef enum {
+    WIFI_NONE_SLEEP = 0, WIFI_LIGHT_SLEEP = 2, WIFI_MODEM_SLEEP = 3
+} WiFiSleepType_t;
+
+
+extern "C" {
+typedef STAILQ_HEAD(, bss_info)
+bss_info_head_t;
+}
+
+#endif /* ESP8266WIFITYPE_H_ */
